Move connection/channel handshake implementations to ProvideCosmosPayloadTypes
