--- crates/cosmos/cosmos-client-components/src/impls/types/connection_handshake_payload.rs
@@ -1,23 +1,0 @@
-use cgp_core::Async;
-use hermes_relayer_components::chain::traits::types::connection::ProvideConnectionHandshakePayloadTypes;
-
-use crate::types::payloads::connection::{
-    CosmosConnectionOpenAckPayload, CosmosConnectionOpenConfirmPayload,
-    CosmosConnectionOpenInitPayload, CosmosConnectionOpenTryPayload,
-};
-
-pub struct ProvideCosmosConnectionHandshakePayloads;
-
-impl<Chain, Counterparty> ProvideConnectionHandshakePayloadTypes<Chain, Counterparty>
-    for ProvideCosmosConnectionHandshakePayloads
-where
-    Chain: Async,
-{
-    type ConnectionOpenInitPayload = CosmosConnectionOpenInitPayload;
-
-    type ConnectionOpenTryPayload = CosmosConnectionOpenTryPayload;
-
-    type ConnectionOpenAckPayload = CosmosConnectionOpenAckPayload;
-
-    type ConnectionOpenConfirmPayload = CosmosConnectionOpenConfirmPayload;
-}
