+++ crates/cosmos/cosmos-client-components/src/impls/types/mod.rs
@@ -1,5 +1,3 @@
 pub mod chain;
-pub mod channel_handshake_payload;
-pub mod connection_handshake_payload;
 pub mod create_client_options;
 pub mod payload;
