+++ crates/cosmos/cosmos-client-components/src/impls/types/payload.rs
@@ -1,9 +1,44 @@
 use cgp_core::Async;
+use hermes_relayer_components::chain::traits::types::channel::ProvideChannelHandshakePayloadTypes;
+use hermes_relayer_components::chain::traits::types::connection::ProvideConnectionHandshakePayloadTypes;
 use hermes_relayer_components::chain::traits::types::packets::receive::ProvideReceivePacketPayloadType;
 
+use crate::types::payloads::channel::{
+    CosmosChannelOpenAckPayload, CosmosChannelOpenConfirmPayload, CosmosChannelOpenTryPayload,
+};
+use crate::types::payloads::connection::{
+    CosmosConnectionOpenAckPayload, CosmosConnectionOpenConfirmPayload,
+    CosmosConnectionOpenInitPayload, CosmosConnectionOpenTryPayload,
+};
 use crate::types::payloads::packet::CosmosReceivePacketPayload;
 
 pub struct ProvideCosmosPayloadTypes;
+
+impl<Chain, Counterparty> ProvideChannelHandshakePayloadTypes<Chain, Counterparty>
+    for ProvideCosmosPayloadTypes
+where
+    Chain: Async,
+{
+    type ChannelOpenTryPayload = CosmosChannelOpenTryPayload;
+
+    type ChannelOpenAckPayload = CosmosChannelOpenAckPayload;
+
+    type ChannelOpenConfirmPayload = CosmosChannelOpenConfirmPayload;
+}
+
+impl<Chain, Counterparty> ProvideConnectionHandshakePayloadTypes<Chain, Counterparty>
+    for ProvideCosmosPayloadTypes
+where
+    Chain: Async,
+{
+    type ConnectionOpenInitPayload = CosmosConnectionOpenInitPayload;
+
+    type ConnectionOpenTryPayload = CosmosConnectionOpenTryPayload;
+
+    type ConnectionOpenAckPayload = CosmosConnectionOpenAckPayload;
+
+    type ConnectionOpenConfirmPayload = CosmosConnectionOpenConfirmPayload;
+}
 
 impl<Chain, Counterparty> ProvideReceivePacketPayloadType<Chain, Counterparty>
     for ProvideCosmosPayloadTypes
